Add files via upload
It is an honor for us to have the privilege to invite you to “IMTEX Exhibition 2023”, International Machine Tool & Manufacturing Technology Exhibition. The Venue for the exhibition is BIEC Bangalore, India. The Show will be held on 19th -25th of January 2023. The Timing for the exhibition is from 10 AM to 6 PM. 

Click here to get more information: https://www.seimitsu.in/contact.html

Reference link: https://www.seimitsu.in/News-IMTEX-2023.html